integration-tests/requirements.txt
pyyaml==6.0.2
pytest==8.3.3
jsonschema==4.23.0
ruff==0.14.7
